fix: status updates creates infinite update loop
The operator is triggered every time we patch the `update` field, which results in an infinite update cycle.

Replace the `status` field with an `Event` object that represents the current status and includes info about the last operation: `Applying`, `OK` or an error.

(refactor some of the hook code so it's easier to test. bare in mind that `hook.test.js` doesn't really work out of the box at the moment).