feat: report schema errors from synthesis
We will now report an error when the of a custom resource has changed, and the existing resources do not adhere to the schema.

This helps drift between resources that are already created and the source schema.

We use `XxxSpec.fromJson` to deserialize. Currently this is requires that by convention the name of the "Spec" struct will be `XxxSpec`. But that's something we can fix in the future.